% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{processTask}
\alias{processTask}
\title{Process a task}
\usage{
processTask(task, redis)
}
\arguments{
\item{task}{A Redis key containing the future to process.}

\item{redis}{A hiredis connection.}
}
\description{
Process a task
}
\keyword{internal}
