% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redis.R
\name{redis}
\alias{redis}
\title{Redis-based futures}
\usage{
redis(..., queue = "RJOBS", config = redis_config(), output_queue = NA)
}
\arguments{
\item{...}{Additional named elements of the future.}

\item{queue}{Redis key name of the task queue (Redis list)}

\item{config}{Redis config}

\item{output_queue}{(optional) Redis key name of the work output queue (Redis list)}
}
\value{
An object of class [RedisFuture].
}
\description{
Use the Redis key/value database to define partially fault-tolerant,
asynchronous task queues for elastic distributed computing.
}
