% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedisFuture-class.R
\name{RedisFuture}
\alias{RedisFuture}
\title{A Redis-based future task queue implementation}
\usage{
RedisFuture(
  expr = NULL,
  substitute = TRUE,
  envir = parent.frame(),
  globals = TRUE,
  packages = NULL,
  lazy = FALSE,
  queue = getOption("future.redis.queue", "{{session}}"),
  config = redis_config(),
  output_queue = NA_character_,
  max_retries = 3L,
  ...
)
}
\arguments{
\item{expr}{An \R \link[base]{expression}.}

\item{substitute}{If TRUE, argument \code{expr} is
\code{\link[base]{substitute}()}:ed, otherwise not.}

\item{envir}{The \link{environment} from where global objects should be
identified.}

\item{globals}{(optional) a logical, a character vector, or a named list
to control how globals are handled.
For details, see section 'Globals used by future expressions'
in the help for \code{\link[future:future]{future()}}.}

\item{packages}{(optional) a character vector specifying packages
to be attached in the \R environment evaluating the future.}

\item{lazy}{If FALSE (default), the future is resolved
eagerly (starting immediately), otherwise not.}

\item{queue}{A Redis key name of the task queue, or a}

\item{config}{A \link[redux:redis_config]{redux::redis_config} Redis configuration object.}

\item{output_queue}{(optional) Redis key name of the work output queue
(note: reserved for future use).}

\item{max_retries}{Maximum number of times the future can be re-submitted
to the task queue in the event of failure.}

\item{...}{Additional named elements of the future.}
}
\value{
An object of class \code{RedisFuture}.
}
\description{
Set up the future parameters.
}
\keyword{internal}
