% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedisFuture-class.R
\name{resolved.RedisFuture}
\alias{resolved.RedisFuture}
\title{Check on the status of a future task.}
\usage{
\method{resolved}{RedisFuture}(x, ...)
}
\value{
boolean indicating the task is finished (TRUE) or not (FALSE)
}
\description{
Check on the status of a future task.
}
\keyword{internal}
