% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedisFuture-class.R
\name{result.RedisFuture}
\alias{result.RedisFuture}
\title{Obtain and return a future task result (blocking)}
\usage{
\method{result}{RedisFuture}(future, ...)
}
\arguments{
\item{future}{an object of class \code{Redis.future}}

\item{...}{additional parameters for \code{future} class methods}
}
\description{
Obtain and return a future task result (blocking)
}
