% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils,redis.R
\name{setAlive}
\alias{setAlive}
\title{Start a task liveness thread}
\usage{
setAlive(port, host, key, password)
}
\arguments{
\item{port}{Redis port}

\item{host}{Redis host name}

\item{key}{The task liveness key to maintain}

\item{password}{(optional) Redis password}
}
\value{
Invoked for the side-effect of maintaining a task liveness
key, \code{NULL} is invisibly returned.
}
\description{
\code{setAlive()} and \code{delAlive()} support worker fault tolerance and are only
to be use internally by the package. It's safe to call \code{setAlive()} and
\code{delAlive()} multiple times; at most one copy of the liveness thread is run.
}
\keyword{internal}
