% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{worker}
\alias{worker}
\title{Start a Redis worker process
The worker process blocks for tasks on the specified queue (a Redis list).
The worker continues to process tasks from the queue untile the queue liveness key is removed,
or until a Redis communication or other error occurs; after which the worker exits (quits R).}
\usage{
worker(queue = "RJOBS", linger = 10, config = redis_config(), quit = FALSE)
}
\arguments{
\item{queue}{The Redis task queue}

\item{linger}{in seconds, max time before system checks (including termination).}

\item{config}{Redis configuration (see XXX TODO: add reference)}

\item{quit}{if TRUE, quit R on exit}
}
\description{
Start a Redis worker process
The worker process blocks for tasks on the specified queue (a Redis list).
The worker continues to process tasks from the queue untile the queue liveness key is removed,
or until a Redis communication or other error occurs; after which the worker exits (quits R).
}
