% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{processTask}
\alias{processTask}
\title{Process a task}
\usage{
processTask(task, redis)
}
\arguments{
\item{redis}{A \link[redux:redis_api]{redux::redis_api} object as returned by \code{\link[redux:hiredis]{redux::hiredis()}}
connection.}

\item{queue}{A Redis key containing the future to process.}
}
\description{
Process a task
}
\keyword{internal}
