% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedisFuture-class.R
\name{RedisFuture}
\alias{RedisFuture}
\title{A Redis-based future task queue implementation}
\usage{
RedisFuture(
  expr = NULL,
  substitute = TRUE,
  globals = TRUE,
  packages = NULL,
  envir = parent.frame(),
  lazy = TRUE,
  queue = "RJOBS",
  config = redis_config(),
  output_queue = NA,
  max_retries = 3,
  ...
)
}
\arguments{
\item{expr}{An \R \link[base]{expression}.}

\item{substitute}{If TRUE, argument \code{expr} is
\code{\link[base]{substitute}()}:ed, otherwise not.}

\item{globals}{(optional) a logical, a character vector, a named list, or
a [globals::Globals] object.  If `TRUE`, globals are identified by code
inspection based on `expr` and `tweak` searching from environment
`envir`.  If `FALSE`, no globals are used.  If a character vector, then
globals are identified by lookup based their names `globals` searching
from environment `envir`.  If a named list or a Globals object, the
globals are used as is.}

\item{packages}{(optional) R packages to load on worker processes}

\item{envir}{The \link{environment} from where global objects should be
identified.}

\item{lazy}{If FALSE (default), the future is resolved
eagerly (starting immediately), otherwise not.}

\item{queue}{Redis key name of the task queue (Redis list)}

\item{config}{Redis config}

\item{output_queue}{(optional) Redis key name of the work output queue
(note: reserved for future use).}

\item{max_retries}{Maximum number of times the future can be submitted
to the task queue in the event of failure.}

\item{...}{Additional named elements of the future.}
}
\value{
An object of class `RedisFuture`.
}
\description{
Set up the future parameters.
}
\keyword{internal}
