% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedisFuture-class.R
\name{result.RedisFuture}
\alias{result.RedisFuture}
\title{Obtain and return a future task result (blocking)}
\usage{
\method{result}{RedisFuture}(future, ...)
}
\description{
Obtain and return a future task result (blocking)
}
