% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{startLocalWorkers}
\alias{startLocalWorkers}
\title{Start one or more background R worker processes on the local system.}
\usage{
startLocalWorkers(
  n,
  queue = getOption("future.redis.queue", "{{session}}"),
  config = redis_config(),
  iter = Inf,
  linger = 10,
  log = NULL,
  Rbin = paste(R.home(component = "bin"), "R", sep = "/")
)
}
\arguments{
\item{n}{number of workers to start.}

\item{queue}{Redis task queue name.}

\item{config}{Redis configuration (see \code{\link{redis_config}})).}

\item{iter}{Maximum number of tasks to acquire before exiting.}

\item{linger}{in seconds, max time before system checks (including termination).}

\item{log}{divert stdout and messages to log file.}

\item{Rbin}{full path to the command-line R program.}
}
\value{
NULL is invisibly returned.
}
\description{
Use \code{startLocalWorkers} to start one or more future.redis R worker
processes in the background. The worker processes are started on the local
system using the \code{\link{worker}} function.
}
\details{
Running workers self-terminate after a \code{linger} interval if their task
queue is deleted with the \code{removeQ} function, or if network
communication with the Redis server encounters an error.
}
\examples{
if (redux::redis_available()) {
## The example assumes that a Redis server is running on the local host
## and standard port.

# Register the redis plan on a specified task queue:
plan(redis, queue = "R jobs")

# Start some local R worker processes:
startLocalWorkers(n=2, queue="R jobs", linger=1)

# Alternatively, use the following to run the workers quietly without
# showing their output as they run:
# startLocalWorkers(n=2, queue="R jobs", linger=1, log="/dev/null")

# A function that returns a future (note the scope of N)
f <- \() future({4 * sum((runif(N) ^ 2 + runif(N) ^ 2) < 1) / N}, seed = TRUE)

# Run a simple sampling approximation of pi in parallel using  M * N points:
N <- 1e6  # samples per worker
M <- 10   # iterations
Reduce(sum, Map(value, replicate(M, f()))) / M

# Clean up
removeQ("R jobs")
}
}
\seealso{
\code{\link{redis_config}}, \code{\link{worker}}, \code{\link{removeQ}}
}
